+++ src/ui/ui_revoke_mn_dlg.ui.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-# Form implementation generated from reading ui file '/Users/blogin/Projects/dash-masternode-tool/src/ui/ui_revoke_mn_dlg.ui.autosave'
+# Form implementation generated from reading ui file '/Users/blogin/Projects/absolute-masternode-tool/src/ui/ui_revoke_mn_dlg.ui.autosave'
 #
 # Created by: PyQt5 UI code generator 5.9.2
 #