+++ src/upd_mn_registrar_dlg.py
@@ -218,7 +218,7 @@ class UpdMnRegistrarDlg(QDialog, ui_upd_mn_registrar_dlg.Ui_UpdMnRegistrarDlg, W
         self.lblMessage.setText('<span style="color:#ff6600">By clicking <span style="font-weight:800">&lt;Send '
                                 'Update Transaction&gt;</span> you agree to send the owner private key to '
                                 'the remote RPC node which is necessary to automatically execute the required command ('
-                                'read notes <a href="https://github.com/Bertrand256/dash-masternode-tool/blob/'
+                                'read notes <a href="https://github.com/absolute-community/absolute-masternode-tool/blob/'
                                 'master/doc/deterministic-mn-migration.md#automatic-method-using-public-rpc-'
                                 'nodes-m1">here</a>). If you do not agree, follow the manual steps.</span>')
         self.minimize_dialog_height()
@@ -347,7 +347,7 @@ class UpdMnRegistrarDlg(QDialog, ui_upd_mn_registrar_dlg.Ui_UpdMnRegistrarDlg, W
                 msg = "<ol>" \
                       "<li>Start a Dash Core wallet with sufficient funds to cover a transaction fee.</li>"
                 msg += "<li>Import the owner private key into the Dash Core wallet if you haven't done this " \
-                       "before (<a href=\"https://github.com/Bertrand256/dash-masternode-tool/blob/master/doc/" \
+                       "before (<a href=\"https://github.com/absolute-community/absolute-masternode-tool/blob/master/doc/" \
                        "deterministic-mn-migration.md#can-i-modify-the-payout-address-without-resetting-the-" \
                        "place-in-the-payment-queue\">details</a>).</li>"
                 msg += "<li>Execute the following command in the Dash Core debug console:<br><br>"